requirements.txt
openai<1.0.0
flask
flask-cors
python-dotenv
pandas
